refactor: Move components into different files
